requirements.txt
Django>=2.1.0
djangorestframework>=3.10.0
